requirements.txt
jsonschema
requests
behave >= 1.2.6
PyHamcrest >= 1.9
traceback2 >= 1.4.0
parse >= 1.8.2
parse_type >= 0.4.2
py >= 1.4.17
